+++ node/resolvers/Routes/index.ts
@@ -1,4 +1,4 @@
-import { ForbiddenError } from '@vtex/api'
+import { UserInputError, ForbiddenError } from '@vtex/api'
 
 import {
   COST_CENTER_DATA_ENTITY,
@@ -70,11 +70,102 @@ const Index = {
         route: {
           params: { orderId },
         },
+        storeUserAuthToken,
+        sessionToken,
+        logger,
       },
-      clients: { oms },
+      clients: { vtexId, session, oms, storefrontPermissions },
     } = ctx
 
+    if (!orderId) {
+      throw new UserInputError('Order ID is required')
+    }
+
+    const token: any = storeUserAuthToken
+
+    if (!token) {
+      throw new ForbiddenError('Access denied')
+    }
+
+    const authUser = await vtexId.getAuthenticatedUser(token)
+
+    const sessionData = await session
+      .getSession(sessionToken as string, ['*'])
+      .then((currentSession: any) => {
+        return currentSession.sessionData
+      })
+      .catch((error: any) => {
+        logger.error({
+          error,
+          message: 'orders-getSession-error',
+        })
+
+        return null
+      })
+
+    const profileEmail = sessionData?.namespaces?.profile?.email?.value
+
+    const checkPermissionAgainstOrder = (
+      permissions: string[],
+      orderData: any
+    ) => {
+      if (
+        authUser?.user === orderData?.clientProfileData?.email ||
+        profileEmail === orderData?.clientProfileData?.email
+      ) {
+        return true
+      }
+
+      if (permissions?.includes('all-orders')) {
+        return true
+      }
+
+      if (
+        permissions?.includes('organization-orders') &&
+        sessionData?.namespaces['storefront-permissions']?.organization
+          ?.value === orderData?.marketingData?.utmCampaign
+      ) {
+        return true
+      }
+
+      if (
+        permissions?.includes('costcenter-orders') &&
+        sessionData?.namespaces['storefront-permissions']?.costcenter?.value ===
+          orderData?.marketingData?.utmMedium
+      ) {
+        return true
+      }
+
+      return false
+    }
+
+    const {
+      data: { checkUserPermission },
+    }: any = await storefrontPermissions
+      .checkUserPermission('vtex.b2b-orders-history@0.x')
+      .catch((error: any) => {
+        logger.error({
+          message: 'checkUserPermission-error',
+          error,
+        })
+
+        return {
+          data: {
+            checkUserPermission: null,
+          },
+        }
+      })
+
     const order: any = await oms.order(String(orderId))
+
+    const permitted = checkPermissionAgainstOrder(
+      checkUserPermission?.permissions,
+      order
+    )
+
+    if (!permitted) {
+      throw new ForbiddenError('Access denied')
+    }
 
     ctx.set('Content-Type', 'application/json')
     ctx.set('Cache-Control', 'no-cache, no-store')
@@ -112,15 +203,15 @@ const Index = {
       })
 
     const filterByPermission = (permissions: string[]) => {
-      if (permissions.indexOf('all-orders') !== -1) {
+      if (permissions.includes('all-orders')) {
         return ``
       }
 
-      if (permissions.indexOf('organization-orders') !== -1) {
+      if (permissions.includes('organization-orders')) {
         return `&f_UtmCampaign=${sessionData.namespaces['storefront-permissions'].organization.value}`
       }
 
-      if (permissions.indexOf('costcenter-orders') !== -1) {
+      if (permissions.includes('costcenter-orders')) {
         return `&f_UtmMedium=${sessionData.namespaces['storefront-permissions'].costcenter.value}`
       }
 
@@ -167,6 +258,36 @@ const Index = {
     ctx.response.body = orders
 
     ctx.response.status = 200
+  },
+  requestCancellation: async (ctx: Context) => {
+    const {
+      vtex: {
+        route: {
+          params: { orderId },
+        },
+      },
+      clients: { checkout },
+    } = ctx
+
+    if (!orderId) {
+      throw new UserInputError('Order ID is required')
+    }
+
+    // use order route handler to verify user has access to order
+    Index.order(ctx).catch(() => {
+      throw new ForbiddenError('Access denied')
+    })
+
+    const requestCancellationResponse = await checkout.requestCancellation(
+      orderId as string
+    )
+
+    ctx.set('Content-Type', 'application/json')
+    ctx.set('Cache-Control', 'no-cache, no-store')
+
+    ctx.response.body = requestCancellationResponse.data
+
+    ctx.response.status = requestCancellationResponse.status
   },
 }